chore: Move repository tests to test/unit/App/Repository
