+++ tests/unit/App/Repository/ContactRepositoryTest.php
@@ -9,7 +9,7 @@
  * @license    https://www.gnu.org/licenses/agpl-3.0.html GNU Affero General Public License Version 3 (AGPL3)
  */
 
-namespace Tests\Unit\Ushahidi;
+namespace Tests\Unit\App\Repository;
 
 use Ushahidi_Repository_Contact;
 use Database_MySQLi;
